add toast component (#4826)
* move tamagui config to ui package

* add toast

* add toast confirmation on post forward

* revert NotebookPost changes
