+++ cosmos.imports.ts
@@ -7,66 +7,68 @@ import * as fixture0 from './packages/app/fixtures/ViewReactionsSheet.fixture';
 import * as fixture1 from './packages/app/fixtures/VideoEmbed.fixture';
 import * as fixture2 from './packages/app/fixtures/UserProfileScreen.fixture';
 import * as fixture3 from './packages/app/fixtures/TwitterVerification.fixture';
-import * as fixture4 from './packages/app/fixtures/Text.fixture';
-import * as fixture5 from './packages/app/fixtures/SummaryCollectionView.fixture';
-import * as fixture6 from './packages/app/fixtures/SplashSequence.fixture';
-import * as fixture7 from './packages/app/fixtures/SearchBar.fixture';
-import * as fixture8 from './packages/app/fixtures/ScreenHeader.fixture';
-import * as fixture9 from './packages/app/fixtures/ReferenceSkeleton.fixture';
-import * as fixture10 from './packages/app/fixtures/ProfileSheet.fixture';
-import * as fixture11 from './packages/app/fixtures/ProfileBlock.fixture';
-import * as fixture12 from './packages/app/fixtures/Pressable.fixture';
-import * as fixture13 from './packages/app/fixtures/PostScreen.fixture';
-import * as fixture14 from './packages/app/fixtures/PostReference.fixture';
-import * as fixture15 from './packages/app/fixtures/ParentAgnosticKeyboardAvoidingView.fixture';
-import * as fixture16 from './packages/app/fixtures/MetaEditorScreen.fixture';
-import * as fixture17 from './packages/app/fixtures/MessageInput.fixture';
-import * as fixture18 from './packages/app/fixtures/MessageActions.fixture';
-import * as fixture19 from './packages/app/fixtures/InviteUsersSheet.fixture';
-import * as fixture20 from './packages/app/fixtures/Input.fixture';
-import * as fixture21 from './packages/app/fixtures/ImageViewer.fixture';
-import * as fixture22 from './packages/app/fixtures/GroupListItem.fixture';
-import * as fixture23 from './packages/app/fixtures/GalleryPost.fixture';
-import * as fixture24 from './packages/app/fixtures/Form.fixture';
-import * as fixture25 from './packages/app/fixtures/EmbedContent.fixture';
-import * as fixture26 from './packages/app/fixtures/EditProfileScreen.fixture';
-import * as fixture27 from './packages/app/fixtures/CreateChatSheet.fixture';
-import * as fixture28 from './packages/app/fixtures/ContactList.fixture';
-import * as fixture29 from './packages/app/fixtures/ChatMessage.fixture';
-import * as fixture31 from './packages/app/fixtures/ChannelHeader.fixture';
-import * as fixture32 from './packages/app/fixtures/ChannelDivider.fixture';
-import * as fixture33 from './packages/app/fixtures/ChannelConfigurationBar.fixture';
-import * as fixture34 from './packages/app/fixtures/Channel.fixture';
-import * as fixture35 from './packages/app/fixtures/Carousel.fixture';
-import * as fixture36 from './packages/app/fixtures/Button.fixture';
-import * as fixture37 from './packages/app/fixtures/BlockSectionList.fixture';
-import * as fixture38 from './packages/app/fixtures/Avatar.fixture';
-import * as fixture39 from './packages/app/fixtures/AudioEmbed.fixture';
-import * as fixture40 from './packages/app/fixtures/AttachmentPreviewList.fixture';
-import * as fixture41 from './packages/app/fixtures/Activity.fixture';
-import * as fixture42 from './packages/app/fixtures/DetailView/NotebookDetailView.fixture';
-import * as fixture43 from './packages/app/fixtures/DetailView/GalleryDetailView.fixture';
-import * as fixture44 from './packages/app/fixtures/DetailView/ChatDetailView.fixture';
-import * as fixture45 from './packages/app/fixtures/ActionSheet/SendPostRetrySheet.fixture';
-import * as fixture46 from './packages/app/fixtures/ActionSheet/ProfileSheet.fixture';
-import * as fixture47 from './packages/app/fixtures/ActionSheet/GroupPreviewSheet.fixture';
-import * as fixture48 from './packages/app/fixtures/ActionSheet/GroupJoinRequestSheet.fixture';
-import * as fixture49 from './packages/app/fixtures/ActionSheet/GenericActionSheet.fixture';
-import * as fixture50 from './packages/app/fixtures/ActionSheet/EditSectionNameSheet.fixture';
-import * as fixture51 from './packages/app/fixtures/ActionSheet/DeleteSheet.fixture';
-import * as fixture52 from './packages/app/fixtures/ActionSheet/CreateChannelSheet.fixture';
-import * as fixture53 from './packages/app/fixtures/ActionSheet/AttachmentSheet.fixture';
-import * as fixture54 from './packages/app/fixtures/ActionSheet/AddGalleryPostSheet.fixture';
-import * as fixture55 from './apps/tlon-mobile/src/App.fixture';
-import * as fixture56 from './apps/tlon-mobile/src/fixtures/SetNicknameScreen.fixture';
-import * as fixture57 from './apps/tlon-mobile/src/fixtures/Onboarding.fixture';
-import * as fixture58 from './apps/tlon-mobile/src/fixtures/InputToolbar.fixture';
+import * as fixture4 from './packages/app/fixtures/Toast.fixture';
+import * as fixture5 from './packages/app/fixtures/Text.fixture';
+import * as fixture6 from './packages/app/fixtures/SummaryCollectionView.fixture';
+import * as fixture7 from './packages/app/fixtures/SplashSequence.fixture';
+import * as fixture8 from './packages/app/fixtures/SearchBar.fixture';
+import * as fixture9 from './packages/app/fixtures/ScreenHeader.fixture';
+import * as fixture10 from './packages/app/fixtures/ReferenceSkeleton.fixture';
+import * as fixture11 from './packages/app/fixtures/ProfileSheet.fixture';
+import * as fixture12 from './packages/app/fixtures/ProfileBlock.fixture';
+import * as fixture13 from './packages/app/fixtures/Pressable.fixture';
+import * as fixture14 from './packages/app/fixtures/PostScreen.fixture';
+import * as fixture15 from './packages/app/fixtures/PostReference.fixture';
+import * as fixture16 from './packages/app/fixtures/ParentAgnosticKeyboardAvoidingView.fixture';
+import * as fixture17 from './packages/app/fixtures/MetaEditorScreen.fixture';
+import * as fixture18 from './packages/app/fixtures/MessageInput.fixture';
+import * as fixture19 from './packages/app/fixtures/MessageActions.fixture';
+import * as fixture20 from './packages/app/fixtures/InviteUsersSheet.fixture';
+import * as fixture21 from './packages/app/fixtures/Input.fixture';
+import * as fixture22 from './packages/app/fixtures/ImageViewer.fixture';
+import * as fixture23 from './packages/app/fixtures/GroupListItem.fixture';
+import * as fixture24 from './packages/app/fixtures/GalleryPost.fixture';
+import * as fixture25 from './packages/app/fixtures/Form.fixture';
+import * as fixture26 from './packages/app/fixtures/EmbedContent.fixture';
+import * as fixture27 from './packages/app/fixtures/EditProfileScreen.fixture';
+import * as fixture28 from './packages/app/fixtures/CreateChatSheet.fixture';
+import * as fixture29 from './packages/app/fixtures/ContactList.fixture';
+import * as fixture30 from './packages/app/fixtures/ChatOptionsSheet.fixture';
+import * as fixture31 from './packages/app/fixtures/ChatMessage.fixture';
+import * as fixture32 from './packages/app/fixtures/ChannelHeader.fixture';
+import * as fixture33 from './packages/app/fixtures/ChannelDivider.fixture';
+import * as fixture34 from './packages/app/fixtures/ChannelConfigurationBar.fixture';
+import * as fixture35 from './packages/app/fixtures/Channel.fixture';
+import * as fixture36 from './packages/app/fixtures/Carousel.fixture';
+import * as fixture37 from './packages/app/fixtures/Button.fixture';
+import * as fixture38 from './packages/app/fixtures/BlockSectionList.fixture';
+import * as fixture39 from './packages/app/fixtures/Avatar.fixture';
+import * as fixture40 from './packages/app/fixtures/AudioEmbed.fixture';
+import * as fixture41 from './packages/app/fixtures/AttachmentPreviewList.fixture';
+import * as fixture42 from './packages/app/fixtures/Activity.fixture';
+import * as fixture43 from './packages/app/fixtures/DetailView/NotebookDetailView.fixture';
+import * as fixture44 from './packages/app/fixtures/DetailView/GalleryDetailView.fixture';
+import * as fixture45 from './packages/app/fixtures/DetailView/ChatDetailView.fixture';
+import * as fixture46 from './packages/app/fixtures/ActionSheet/SendPostRetrySheet.fixture';
+import * as fixture47 from './packages/app/fixtures/ActionSheet/ProfileSheet.fixture';
+import * as fixture48 from './packages/app/fixtures/ActionSheet/GroupPreviewSheet.fixture';
+import * as fixture49 from './packages/app/fixtures/ActionSheet/GroupJoinRequestSheet.fixture';
+import * as fixture50 from './packages/app/fixtures/ActionSheet/GenericActionSheet.fixture';
+import * as fixture51 from './packages/app/fixtures/ActionSheet/EditSectionNameSheet.fixture';
+import * as fixture52 from './packages/app/fixtures/ActionSheet/DeleteSheet.fixture';
+import * as fixture53 from './packages/app/fixtures/ActionSheet/CreateChannelSheet.fixture';
+import * as fixture54 from './packages/app/fixtures/ActionSheet/AttachmentSheet.fixture';
+import * as fixture55 from './packages/app/fixtures/ActionSheet/AddGalleryPostSheet.fixture';
+import * as fixture56 from './apps/tlon-mobile/src/App.fixture';
+import * as fixture57 from './apps/tlon-mobile/src/fixtures/SetNicknameScreen.fixture';
+import * as fixture58 from './apps/tlon-mobile/src/fixtures/Onboarding.fixture';
+import * as fixture59 from './apps/tlon-mobile/src/fixtures/InputToolbar.fixture';
 
 import * as decorator0 from './packages/app/fixtures/cosmos.decorator';
 import * as decorator1 from './apps/tlon-mobile/src/fixtures/cosmos.decorator';
 
 export const rendererConfig: RendererConfig = {
-  "playgroundUrl": "http://localhost:5001",
+  "playgroundUrl": "http://localhost:5002",
   "rendererUrl": null
 };
 
@@ -75,60 +77,62 @@ const fixtures = {
   'packages/app/fixtures/VideoEmbed.fixture.tsx': { module: fixture1 },
   'packages/app/fixtures/UserProfileScreen.fixture.tsx': { module: fixture2 },
   'packages/app/fixtures/TwitterVerification.fixture.tsx': { module: fixture3 },
-  'packages/app/fixtures/Text.fixture.tsx': { module: fixture4 },
-  'packages/app/fixtures/SummaryCollectionView.fixture.tsx': { module: fixture5 },
-  'packages/app/fixtures/SplashSequence.fixture.tsx': { module: fixture6 },
-  'packages/app/fixtures/SearchBar.fixture.tsx': { module: fixture7 },
-  'packages/app/fixtures/ScreenHeader.fixture.tsx': { module: fixture8 },
-  'packages/app/fixtures/ReferenceSkeleton.fixture.tsx': { module: fixture9 },
-  'packages/app/fixtures/ProfileSheet.fixture.tsx': { module: fixture10 },
-  'packages/app/fixtures/ProfileBlock.fixture.tsx': { module: fixture11 },
-  'packages/app/fixtures/Pressable.fixture.tsx': { module: fixture12 },
-  'packages/app/fixtures/PostScreen.fixture.tsx': { module: fixture13 },
-  'packages/app/fixtures/PostReference.fixture.tsx': { module: fixture14 },
-  'packages/app/fixtures/ParentAgnosticKeyboardAvoidingView.fixture.tsx': { module: fixture15 },
-  'packages/app/fixtures/MetaEditorScreen.fixture.tsx': { module: fixture16 },
-  'packages/app/fixtures/MessageInput.fixture.tsx': { module: fixture17 },
-  'packages/app/fixtures/MessageActions.fixture.tsx': { module: fixture18 },
-  'packages/app/fixtures/InviteUsersSheet.fixture.tsx': { module: fixture19 },
-  'packages/app/fixtures/Input.fixture.tsx': { module: fixture20 },
-  'packages/app/fixtures/ImageViewer.fixture.tsx': { module: fixture21 },
-  'packages/app/fixtures/GroupListItem.fixture.tsx': { module: fixture22 },
-  'packages/app/fixtures/GalleryPost.fixture.tsx': { module: fixture23 },
-  'packages/app/fixtures/Form.fixture.tsx': { module: fixture24 },
-  'packages/app/fixtures/EmbedContent.fixture.tsx': { module: fixture25 },
-  'packages/app/fixtures/EditProfileScreen.fixture.tsx': { module: fixture26 },
-  'packages/app/fixtures/CreateChatSheet.fixture.tsx': { module: fixture27 },
-  'packages/app/fixtures/ContactList.fixture.tsx': { module: fixture28 },
-  'packages/app/fixtures/ChatMessage.fixture.tsx': { module: fixture29 },
-  'packages/app/fixtures/ChannelHeader.fixture.tsx': { module: fixture31 },
-  'packages/app/fixtures/ChannelDivider.fixture.tsx': { module: fixture32 },
-  'packages/app/fixtures/ChannelConfigurationBar.fixture.tsx': { module: fixture33 },
-  'packages/app/fixtures/Channel.fixture.tsx': { module: fixture34 },
-  'packages/app/fixtures/Carousel.fixture.tsx': { module: fixture35 },
-  'packages/app/fixtures/Button.fixture.tsx': { module: fixture36 },
-  'packages/app/fixtures/BlockSectionList.fixture.tsx': { module: fixture37 },
-  'packages/app/fixtures/Avatar.fixture.tsx': { module: fixture38 },
-  'packages/app/fixtures/AudioEmbed.fixture.tsx': { module: fixture39 },
-  'packages/app/fixtures/AttachmentPreviewList.fixture.tsx': { module: fixture40 },
-  'packages/app/fixtures/Activity.fixture.tsx': { module: fixture41 },
-  'packages/app/fixtures/DetailView/NotebookDetailView.fixture.tsx': { module: fixture42 },
-  'packages/app/fixtures/DetailView/GalleryDetailView.fixture.tsx': { module: fixture43 },
-  'packages/app/fixtures/DetailView/ChatDetailView.fixture.tsx': { module: fixture44 },
-  'packages/app/fixtures/ActionSheet/SendPostRetrySheet.fixture.tsx': { module: fixture45 },
-  'packages/app/fixtures/ActionSheet/ProfileSheet.fixture.tsx': { module: fixture46 },
-  'packages/app/fixtures/ActionSheet/GroupPreviewSheet.fixture.tsx': { module: fixture47 },
-  'packages/app/fixtures/ActionSheet/GroupJoinRequestSheet.fixture.tsx': { module: fixture48 },
-  'packages/app/fixtures/ActionSheet/GenericActionSheet.fixture.tsx': { module: fixture49 },
-  'packages/app/fixtures/ActionSheet/EditSectionNameSheet.fixture.tsx': { module: fixture50 },
-  'packages/app/fixtures/ActionSheet/DeleteSheet.fixture.tsx': { module: fixture51 },
-  'packages/app/fixtures/ActionSheet/CreateChannelSheet.fixture.tsx': { module: fixture52 },
-  'packages/app/fixtures/ActionSheet/AttachmentSheet.fixture.tsx': { module: fixture53 },
-  'packages/app/fixtures/ActionSheet/AddGalleryPostSheet.fixture.tsx': { module: fixture54 },
-  'apps/tlon-mobile/src/App.fixture.tsx': { module: fixture55 },
-  'apps/tlon-mobile/src/fixtures/SetNicknameScreen.fixture.tsx': { module: fixture56 },
-  'apps/tlon-mobile/src/fixtures/Onboarding.fixture.tsx': { module: fixture57 },
-  'apps/tlon-mobile/src/fixtures/InputToolbar.fixture.tsx': { module: fixture58 }
+  'packages/app/fixtures/Toast.fixture.tsx': { module: fixture4 },
+  'packages/app/fixtures/Text.fixture.tsx': { module: fixture5 },
+  'packages/app/fixtures/SummaryCollectionView.fixture.tsx': { module: fixture6 },
+  'packages/app/fixtures/SplashSequence.fixture.tsx': { module: fixture7 },
+  'packages/app/fixtures/SearchBar.fixture.tsx': { module: fixture8 },
+  'packages/app/fixtures/ScreenHeader.fixture.tsx': { module: fixture9 },
+  'packages/app/fixtures/ReferenceSkeleton.fixture.tsx': { module: fixture10 },
+  'packages/app/fixtures/ProfileSheet.fixture.tsx': { module: fixture11 },
+  'packages/app/fixtures/ProfileBlock.fixture.tsx': { module: fixture12 },
+  'packages/app/fixtures/Pressable.fixture.tsx': { module: fixture13 },
+  'packages/app/fixtures/PostScreen.fixture.tsx': { module: fixture14 },
+  'packages/app/fixtures/PostReference.fixture.tsx': { module: fixture15 },
+  'packages/app/fixtures/ParentAgnosticKeyboardAvoidingView.fixture.tsx': { module: fixture16 },
+  'packages/app/fixtures/MetaEditorScreen.fixture.tsx': { module: fixture17 },
+  'packages/app/fixtures/MessageInput.fixture.tsx': { module: fixture18 },
+  'packages/app/fixtures/MessageActions.fixture.tsx': { module: fixture19 },
+  'packages/app/fixtures/InviteUsersSheet.fixture.tsx': { module: fixture20 },
+  'packages/app/fixtures/Input.fixture.tsx': { module: fixture21 },
+  'packages/app/fixtures/ImageViewer.fixture.tsx': { module: fixture22 },
+  'packages/app/fixtures/GroupListItem.fixture.tsx': { module: fixture23 },
+  'packages/app/fixtures/GalleryPost.fixture.tsx': { module: fixture24 },
+  'packages/app/fixtures/Form.fixture.tsx': { module: fixture25 },
+  'packages/app/fixtures/EmbedContent.fixture.tsx': { module: fixture26 },
+  'packages/app/fixtures/EditProfileScreen.fixture.tsx': { module: fixture27 },
+  'packages/app/fixtures/CreateChatSheet.fixture.tsx': { module: fixture28 },
+  'packages/app/fixtures/ContactList.fixture.tsx': { module: fixture29 },
+  'packages/app/fixtures/ChatOptionsSheet.fixture.tsx': { module: fixture30 },
+  'packages/app/fixtures/ChatMessage.fixture.tsx': { module: fixture31 },
+  'packages/app/fixtures/ChannelHeader.fixture.tsx': { module: fixture32 },
+  'packages/app/fixtures/ChannelDivider.fixture.tsx': { module: fixture33 },
+  'packages/app/fixtures/ChannelConfigurationBar.fixture.tsx': { module: fixture34 },
+  'packages/app/fixtures/Channel.fixture.tsx': { module: fixture35 },
+  'packages/app/fixtures/Carousel.fixture.tsx': { module: fixture36 },
+  'packages/app/fixtures/Button.fixture.tsx': { module: fixture37 },
+  'packages/app/fixtures/BlockSectionList.fixture.tsx': { module: fixture38 },
+  'packages/app/fixtures/Avatar.fixture.tsx': { module: fixture39 },
+  'packages/app/fixtures/AudioEmbed.fixture.tsx': { module: fixture40 },
+  'packages/app/fixtures/AttachmentPreviewList.fixture.tsx': { module: fixture41 },
+  'packages/app/fixtures/Activity.fixture.tsx': { module: fixture42 },
+  'packages/app/fixtures/DetailView/NotebookDetailView.fixture.tsx': { module: fixture43 },
+  'packages/app/fixtures/DetailView/GalleryDetailView.fixture.tsx': { module: fixture44 },
+  'packages/app/fixtures/DetailView/ChatDetailView.fixture.tsx': { module: fixture45 },
+  'packages/app/fixtures/ActionSheet/SendPostRetrySheet.fixture.tsx': { module: fixture46 },
+  'packages/app/fixtures/ActionSheet/ProfileSheet.fixture.tsx': { module: fixture47 },
+  'packages/app/fixtures/ActionSheet/GroupPreviewSheet.fixture.tsx': { module: fixture48 },
+  'packages/app/fixtures/ActionSheet/GroupJoinRequestSheet.fixture.tsx': { module: fixture49 },
+  'packages/app/fixtures/ActionSheet/GenericActionSheet.fixture.tsx': { module: fixture50 },
+  'packages/app/fixtures/ActionSheet/EditSectionNameSheet.fixture.tsx': { module: fixture51 },
+  'packages/app/fixtures/ActionSheet/DeleteSheet.fixture.tsx': { module: fixture52 },
+  'packages/app/fixtures/ActionSheet/CreateChannelSheet.fixture.tsx': { module: fixture53 },
+  'packages/app/fixtures/ActionSheet/AttachmentSheet.fixture.tsx': { module: fixture54 },
+  'packages/app/fixtures/ActionSheet/AddGalleryPostSheet.fixture.tsx': { module: fixture55 },
+  'apps/tlon-mobile/src/App.fixture.tsx': { module: fixture56 },
+  'apps/tlon-mobile/src/fixtures/SetNicknameScreen.fixture.tsx': { module: fixture57 },
+  'apps/tlon-mobile/src/fixtures/Onboarding.fixture.tsx': { module: fixture58 },
+  'apps/tlon-mobile/src/fixtures/InputToolbar.fixture.tsx': { module: fixture59 }
 };
 
 const decorators = {
