+++ packages/ui/src/index.ts
@@ -30,4 +30,5 @@ export * from './components/View';
 export * from './contexts/globalSearch';
 export { useCopy } from './hooks/useCopy';
 export { default as useIsWindowNarrow } from './hooks/useIsWindowNarrow';
+export * from './components/Toast';
 export * from './utils';
